NFC: UI freeze on WiFi connect using Tag
If WiFi is disabled and then try to connect to AP
using Tag read, we are not able to click on Menu.

When we tap Tag with DUT to connect to AP, new activity
is created, but it was not finished. As it was coming
on top of current activity user will think UI is freezed.

Issue is coming only when Wi-Fi is disabled before Tag read,
added code to finish activity once Wi-Fi is enabled.

Change-Id: I4caac53730fafcaad4ff48f365fd0295a67ac8ed